more refinement + documentation

+++ app/src/Controllers/VenueController.php
@@ -17,20 +17,24 @@ class VenueController extends BaseController
 {
     public function __construct(private VenueModel $venue_model, private VenuesService $venues_service) {}
 
+    /**
+     * Function to handle get all venues. Has filtering
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @throws \Slim\Exception\HttpNotFoundException
+     * @return \Psr\Http\Message\ResponseInterface
+     */
     public function handleGetVenues(Request $request, Response $response): Response
     {
         // Retrieve the set of parameters.
         $req_params = $request->getQueryParams();
-
         // Overriding the default pagination if specified in the request.
         $current_page = $req_params["current_page"] ?? 1;
         $page_size = $req_params["page_size"] ?? 15;
-
         $this->venue_model->setPaginationOptions($current_page, $page_size);
 
         // Calling model to fetch records
         $venues = $this->venue_model->getVenues($request, $req_params);
-
         if (empty($venues["data"])) {
             throw new HttpNotFoundException(
                 $request,
@@ -43,10 +47,19 @@ class VenueController extends BaseController
         }
     }
 
+    /**
+     * Function to handle getting venue by id
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @param array $uri_args
+     * @throws \Slim\Exception\HttpNotFoundException
+     * @return \Psr\Http\Message\ResponseInterface
+     */
     public function handleGetVenueById(Request $request, Response $response, array $uri_args): Response
     {
+        // Retrieve id from Uri
         $venue_id = $uri_args["venue_id"];
-        // Check if venue id is provided.
+        // Verify if id isset.
         if (!isset($venue_id)) {
             return $this->renderJson(
                 $response,
@@ -55,11 +68,9 @@ class VenueController extends BaseController
                     "code" => "400",
                     "message" => "No venue Id was provided",
                     "hint" => "It must include only numbers between 0-99."
-                ],
-                StatusCodeInterface::STATUS_BAD_REQUEST
+                ],StatusCodeInterface::STATUS_BAD_REQUEST
             );
         }
-
         // Validate the format of the player id
         if (!ValidationHelper::isIntAndInRange($venue_id, 0, 999)) {
             return $this->renderJson(
@@ -69,10 +80,10 @@ class VenueController extends BaseController
                     "code" => "400",
                     "message" => "The venue Id provided is invalid.",
                     "hint" => "It must include only numbers between 0-99."
-                ],
-                StatusCodeInterface::STATUS_BAD_REQUEST
+                ],StatusCodeInterface::STATUS_BAD_REQUEST
             );
         }
+        // Call getVenuesById from the model
         $venue = $this->venue_model->getVenueById($venue_id);
         if ($venue === false) {
             throw new HttpNotFoundException(
@@ -83,12 +94,12 @@ class VenueController extends BaseController
         return $this->renderJson($response, $venue);
     }
 
-
     /**
      * Function that handles the creation and insertion of a new venue
-     * @param Request $request
-     * @param Response $response
-     * @return void
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @throws \App\Exceptions\HttpNoDataProvidedException
+     * @return \Psr\Http\Message\ResponseInterface
      */
     public function handleCreateVenue(Request $request, Response $response): Response
     {
@@ -97,12 +108,11 @@ class VenueController extends BaseController
         if(!isset($data) || empty($data)){
             throw new HttpNoDataProvidedException($request);
         }
-
-        // Create venue using venues service
-        $result = $this->venues_service->CreateVenue($request, $data);
+        // Call createVenue using venues service
+        $result = $this->venues_service->createVenue($request, $data);
+        // Preparing payload
         $status_code = 201;
         if ($result->isSuccess()) {
-            //Prepare success stmt
             $payload["success"] = true;
         } else {
             $status_code = 400;
@@ -115,14 +125,22 @@ class VenueController extends BaseController
         return $this->renderJson($response, $payload, $status_code);
     }
 
-    //? ToBeDebugged
+    /**
+     * Summary of handleUpdateVenue
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @return \Psr\Http\Message\ResponseInterface
+     */
     public function handleUpdateVenue(Request $request, Response $response) : Response
     {
         // Retrieving the venue id to be updated and the update information
         $data = $request->getParsedBody();
-
+        if(!isset($data) || empty($data)){
+            throw new HttpNoDataProvidedException($request);
+        }
+        // Calling updateVenue from service
         $result = $this->venues_service->updateVenue($request, $data);
-
+        // Preparing payload
         $status_code = 201;
         if ($result->isSuccess()){
             $payload["success"] = true;
@@ -137,16 +155,23 @@ class VenueController extends BaseController
         return $this->renderJson($response, $payload, $status_code);
     }
 
+    /**
+     * Summary of handleDeleteVenue
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @return \Psr\Http\Message\ResponseInterface
+     */
     public function handleDeleteVenue(Request $request, Response $response) : Response
     {
-        // Retrieve if for where clause
+        // Retrieve id for where clause
         $data = $request->getParsedBody();
-        $where = ['venue_id' => $data["venue_id"]];
-        // dd($data);
+        if(!isset($data) || empty($data)){
+            throw new HttpNoDataProvidedException($request);
+        }
 
         // Call DeleteVenue for validation and execution
+        $where = ['venue_id' => $data["venue_id"]];
         $result = $this->venues_service->deleteVenue($request, $where);
-
 
         // Process response of deleteVenue operation
         $status_code = 200;
@@ -162,6 +187,12 @@ class VenueController extends BaseController
         return $this->renderJson($response, $payload, $status_code);
     }
 
+    /**
+     * Summary of handleLog
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     * @return \Psr\Http\Message\ResponseInterface
+     */
     public function handleLog(Request $request, Response $response): Response{
 
         echo 'Logging process';
@@ -179,8 +210,6 @@ class VenueController extends BaseController
             $log_record,
             $extra
         );
-
         return $response;
-
     }
 }
